requirements.txt
alembic==1.4.2
aniso8601==8.0.0
Authlib==0.14.1
certifi==2019.11.28
cffi>=1.15.1
chardet==3.0.4
click==7.1.1
cryptography>=42.0.8
Flask==1.1.2
Flask-Admin==1.5.5
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==0.16.1
WTForms==2.2.1
